removing cf utils and adding all content to cf controller
