+++ src/main/java/com/jaypos/maratonouBot/listener/slash/CodeforcesCommandsListener.java
@@ -1,13 +1,10 @@
 package com.jaypos.maratonouBot.listener.slash;
 
 import com.jaypos.maratonouBot.controller.CodeforcesControler;
-import com.jaypos.maratonouBot.utils.CodeforcesUtils;
-import com.jaypos.maratonouBot.utils.Util;
 import com.softawii.curupira.annotations.IArgument;
 import com.softawii.curupira.annotations.ICommand;
 import com.softawii.curupira.annotations.IGroup;
 import net.dv8tion.jda.api.EmbedBuilder;
-import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import ru.covariance.codeforcesapi.CodeforcesApiException;
 
@@ -51,7 +48,7 @@ public class CodeforcesCommandsListener {
             if (event.getOption("gym") != null) {
                 gym = event.getOption("gym").getAsBoolean();
             }
-            List<String> nextContestsMessages = Util.cfUtils.getNextContestsListMessages(gym);
+            List<String> nextContestsMessages = cfController.getNextContestsListMessages(gym);
             String response = "";
             for (int i = 0; i < nextContestsMessages.size(); i++) {
                 response += nextContestsMessages.get(i);
